requirements.txt
Django==4.1.2
psycopg2-binary==2.9.5
djangorestframework==3.14.0
drf-spectacular[sidecar]==0.24.2
Pillow==9.2.0
